Tabs styled-component migration
