+++ src/index.tsx
@@ -34,7 +34,7 @@ export { default as Snackbar } from 'snackbar'
 export { default as Stars } from 'stars'
 export { default as Stepper } from 'stepper'
 export { default as SuccessModal } from 'successModal'
-export { default as Tabs } from 'tabs'
+export { default as Tabs, TabStatus } from 'tabs'
 export { default as Text, TextDisplayType, TextTagType } from 'text'
 export { default as TextField } from 'textField'
 export { default as TimePicker } from 'timePicker'
